revert back to 0.2.3

+++ workflow/tools/replace_readgroup.py
@@ -3,6 +3,7 @@ import os
 import subprocess
 import sys
 import json
+import time
 import glob
 
 """
@@ -22,7 +23,7 @@ with open(task_dict['input'].get('metadata_json'), 'r') as f:
     metadata = json.load(f)
 
 output = {
-    'unaligned_rg_replace_dir': cwd
+    'unaligned_rg_replace_dir': None
 }
 
 RG_map = { 'ID': 'readGroupId',
@@ -39,8 +40,14 @@ RG_map = { 'ID': 'readGroupId',
 # the inputs are BAM
 if input_format == 'BAM':
     files = metadata.get('files')
+    output_dir = os.path.join(cwd, 'unaligned_rg_replace')
+    if not os.path.isdir(output_dir): os.makedirs(output_dir)
+    output['unaligned_rg_replace_dir'] = output_dir
 
     for _file in files:
+        file_path = _file.get('path')
+        file_name = _file.get('fileName')
+
         # get all the rg for the _file
         rg_yaml = set()
         rg_replace = {}
@@ -53,49 +60,31 @@ if input_format == 'BAM':
                     to_update = {key: metadata.get(value)}
                 if key in ['ID', 'LB', 'PL', 'PU', 'PM', 'CN', 'DT'] and rg.get(value) or key == 'PI' and isinstance(rg.get(value), int):
                     to_update = {key: rg.get(value)}
+                # if key == 'DT' and re.match('^[0-9]{4}-[0-9]{2}-[0-9]{2}T[0-9]{2}:[0-9]{2}:[0-9]{2}$', str(rg.get(value))):
+                #     to_update = {key: rg.get(value)}
                 if to_update:
                     rg_replace[rg.get('readGroupIdInFile')].update(to_update)
 
-    # do the replacement for all readGroups
-    for f in glob.glob(os.path.join(unaligned_by_rg_dir, '*.bam')):
-        # retrieve the @RG from BAM header
-        try:
-            header = subprocess.check_output(['samtools', 'view', '-H', args.input_bam])
+        # do the replacement for all readGroups
+        for rg_old, rg_new in rg_replace.items():
+            rg_args = []
+            for key, value in rg_new.items():
+                rg_args.append('RG%s=%s' % (key, value))
 
-        except Exception as e:
-            sys.exit('\n%s: Retrieve BAM header failed: %s' % (e, args.input_bam))
-
-        # get @RG and RGID
-        header_array = header.decode('utf-8').rstrip().split('\n')
-        rg_array = []
-        for line in header_array:
-            if not line.startswith("@RG"): continue
-            rg_array.append(line.rstrip())
-
-        if not len(rg_array) == 1: sys.exit(
-            '\n%s: The input bam should only contain one readgroup ID: %s' % args.input_bam)
-
-        for element in rg_array[0].split("\t"):
-            if not element.startswith('ID'): continue
-            rg_old = element.replace("ID:", "")
-
-        rg_args = []
-        for key, value in rg_replace.get(rg_old):
-            rg_args.append('RG%s=%s' % (key, value))
-
-        try:
-            subprocess.run(['java', '-jar', picard,
-                            'AddOrReplaceReadGroups',
-                            'VALIDATION_STRINGENCY=LENIENT',
-                            'I=%s' % f,
-                            'O=%s' % os.path.join(cwd, os.path.basename(f))] + \
-                            rg_args, check=True)
-        except Exception as e:
-            sys.exit('\n%s: ReplaceReadGroups failed: %s' % (e, f))
+            try:
+                subprocess.run(['java', '-jar', picard,
+                                'AddOrReplaceReadGroups',
+                                'VALIDATION_STRINGENCY=LENIENT',
+                                'I=%s' % os.path.join(unaligned_by_rg_dir, rg_old+'.bam'),
+                                'O=%s' % os.path.join(output_dir, rg_new.get('ID')+'.new.bam')] + \
+                                rg_args, check=True)
+            except Exception as e:
+                sys.exit('\n%s: ReplaceReadGroups failed: %s' % (e, os.path.join(unaligned_by_rg_dir, rg_old+'.bam')))
 
     # delete input bam files at the very last moment
-    for f in glob.glob(os.path.join(unaligned_by_rg_dir, "*.bam")):
-        os.remove(f)
+    if os.path.isdir(unaligned_by_rg_dir):
+        for f in glob.glob(os.path.join(unaligned_by_rg_dir, "*.bam")):
+            os.remove(f)
 
 elif input_format == 'FASTQ':
     pass
